+++ miden-tx/src/host/mod.rs
@@ -1,8 +1,7 @@
 use alloc::{collections::BTreeMap, rc::Rc, string::ToString, vec::Vec};
 
 use miden_lib::transaction::{
-    memory::{MemoryAddress, CURRENT_CONSUMED_NOTE_PTR},
-    TransactionEvent, TransactionKernelError, TransactionTrace,
+    memory::CURRENT_CONSUMED_NOTE_PTR, TransactionEvent, TransactionKernelError, TransactionTrace,
 };
 use miden_objects::{
     accounts::{AccountDelta, AccountId, AccountStorage, AccountStub},
@@ -52,7 +51,7 @@ pub struct TransactionHost<A, T> {
 
     /// The list of notes created while executing a transaction stored as note_ptr |-> note_builder
     /// map.
-    output_notes: BTreeMap<MemoryAddress, OutputNoteBuilder>,
+    output_notes: BTreeMap<usize, OutputNoteBuilder>,
 
     /// Provides a way to get a signature for a message into a transaction
     authenticator: Option<Rc<T>>,
@@ -114,13 +113,15 @@ impl<A: AdviceProvider, T: TransactionAuthenticator> TransactionHost<A, T> {
         process: &S,
     ) -> Result<(), TransactionKernelError> {
         let stack = process.get_stack_state();
+        // # => [aux, note_type, sender_acct_id, tag, note_ptr, RECIPIENT, note_idx]
 
-        let note_ptr: MemoryAddress =
-            stack[4].try_into().map_err(TransactionKernelError::MalformedNotePointer)?;
+        let note_idx: usize = stack[9].as_int() as usize;
+
+        assert_eq!(note_idx, self.output_notes.len(), "note index mismatch");
 
         let note_builder = OutputNoteBuilder::new(stack, &self.adv_provider)?;
 
-        self.output_notes.insert(note_ptr, note_builder);
+        self.output_notes.insert(note_idx, note_builder);
 
         Ok(())
     }
@@ -132,17 +133,20 @@ impl<A: AdviceProvider, T: TransactionAuthenticator> TransactionHost<A, T> {
         &mut self,
         process: &S,
     ) -> Result<(), TransactionKernelError> {
-        let note_ptr: MemoryAddress = process
-            .get_stack_item(4)
-            .try_into()
-            .map_err(TransactionKernelError::MalformedNotePointer)?;
+        let stack = process.get_stack_state();
+        //# => [ASSET, note_ptr, num_of_assets, note_idx]
+
+        let note_idx = stack[6].as_int();
+        assert!(note_idx < self.output_notes.len() as u64);
+        let node_idx = note_idx as usize;
+
         let asset = Asset::try_from(process.get_stack_word(0))
             .map_err(TransactionKernelError::MalformedAsset)?;
 
         let note_builder = self
             .output_notes
-            .get_mut(&note_ptr)
-            .ok_or_else(|| TransactionKernelError::MissingNote(format!("{:?}", &note_ptr)))?;
+            .get_mut(&node_idx)
+            .ok_or_else(|| TransactionKernelError::MissingNote(format!("{:?}", &note_idx)))?;
 
         note_builder.add_asset(asset)?;