+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - [BREAKING] Changed rng to mutable reference in note creation functions (#733).
 - Added transaction IDs to the `Block` struct (#734).
 - [BREAKING] Replaced `ToNullifier` trait with `ToInputNoteCommitments`, which includes the `note_id` for delayed note authentication (#732).
+- Changed note pointer from Memory `note_ptr` to `note_index` (#728)
 
 ## 0.3.0 (2024-05-14)